updated design pattern
