+++ app/server/crawler/strategy.py
@@ -1,189 +1,52 @@
-import abc
+from strategy import HouseRepStrategy
+from strategy import SenateRepStrategy
+from strategy import ExecBranchStrategy
 
-class PoliticianStrategyAbstract(object):
-    __metaclass__ = abc.ABCMeta
+house_rep = HouseRepStrategy()
+senate_rep = SenateRepStrategy()
+exec_branch = ExecBranchStrategy()
 
-    @abc.abstractmethod
-        def search_congress(self):
-    @abc.abstractmethod
-        def construct_profile(self):
 
-class HouseRepStrategy(PoliticianStrategyAbstract):
-    def search_congress(self):
-        name = self._name.split(" ")
-        headers = {"X-API-Key":"KgI2lOueGBFwLYWYsicnT4PSQUblFGDEpfj2Gcdd"}
+class Politician(Celebrity):
+    def __init__(self, name, strategy):
+        self.name= name
+        self.strategy= strategy
+        self.occID = 'politician'
+        self.info = {}
 
-        member_ID= 0
-        x=115
-        while x>102:
-            add_on= str(x)+ "/house/members"
-            url= "https://api.propublica.org/congress/v1/" + add_on
-            response= requests.get(url, headers=headers)
-            data= response.json()
-            results= data["results"]
-            result= results[0]
+    def retrieve_info(self):
+        info= self.strategy.construct_profile()
+        wiki_data = search_wiki(self.name)
+        wiki_desc = wiki_data[2][0]
+        info['bio']= wiki_desc
 
-            first_name= name[0]
-            last_name= name[1]
-            congresses= data["results"]
-            congress= congresses[0]
-            check_found=0
-            members= congress["members"]
-            for member in members:
-                if(member["last_name"]==last_name):
-                    if(member["first_name"]== first_name):
-                        member_ID= member["id"]
-                        check_found=1
-                        break
-            if check_found==1:
-                break
-            x=x-1
-        return member_ID    
-    
-    def construct_profile(self):
-        profile= self._profile
-        headers = {"X-API-Key":"KgI2lOueGBFwLYWYsicnT4PSQUblFGDEpfj2Gcdd"}
-        url= "https://api.propublica.org/congress/v1/" + "members/" + member_ID
-        response= requests.get(url, headers=headers)
-        data= response.json()
-        result= data["results"]
-        member= result[0]
-        profile= {}
-        if(member["middle_name"]):
-            profile['name']= member["first_name"] + " " + member["middle_name"] + " " + member["last_name"]
-        else:
-            profile['name']= member["first_name"] + " " +  member["last_name"]
-        profile['party']= ""
-        if member["current_party"]== "R":
-            profile['party']= "Republican"
-        elif member["current_party"]== "D":
-            profile['party']= "Democrat"
-        else:
-            profile['party']= "Other"
-        roles_list= member["roles"]
-        roles= roles_list[0]
-        profile['title']= "U.S. " + roles["chamber"] + ", " + roles["title"]
-        profile['state']= roles["state"]
-        profile['service_span']= "From " + roles["start_date"]+ " to " + roles["end_date"]
-        profile['website']= member["url"]
-        if(member["twitter_account"]):
-            profile['twitter']= "@" + member["twitter_account"]
-        profile['facebook']= member["facebook_account"]
-        profile['address']= roles["office"]
-        profile['phone']= roles["phone"]
-        committees=[]
-        for committee in roles["committees"]:
-            committees.append(committee["name"])
-        profile['committees']=committees
-        return(profile)
+        return info
 
-class SenateRepStrategy(PoliticianStrategyAbstract):
-    def search_congress(self):
-        new_name = self._name.split()
-        headers = {"X-API-Key":"KgI2lOueGBFwLYWYsicnT4PSQUblFGDEpfj2Gcdd"}
+#Types of Politicians 
+class President(Politician):
+    def __init__(self, name):
+        super(President, self).__init__(name, senate_rep)
 
-        member_ID= 0
-        x=115
-        while x>79:
-            add_on= str(x)+ "/senate/members"
-            url= "https://api.propublica.org/congress/v1/" + add_on
-            response= requests.get(url, headers=headers)
-            data= response.json()
-            results= data["results"]
-            result= results[0]
 
-            first_name= new_name[0]
-            last_name= new_name[1]
-            congresses= data["results"]
-            congress= congresses[0]
-            check_found=0
-            members= congress["members"]
-            for member in members:
-                if(member["last_name"]==last_name):
-                    if(member["first_name"]== first_name):
-                        member_ID= member["id"]
-                        check_found=1
-                        break
-            if check_found==1:
-                break
-            x=x-1
-        return member_ID
-    
-    def construct_profile(self):
-        profile= self._profile
-        headers = {"X-API-Key":"KgI2lOueGBFwLYWYsicnT4PSQUblFGDEpfj2Gcdd"}
-        url= "https://api.propublica.org/congress/v1/" + "members/" + member_ID
-        response= requests.get(url, headers=headers)
-        data= response.json()
-        result= data["results"]
-        member= result[0]
-        profile= {}
-        if(member["middle_name"]):
-            profile['name']= member["first_name"] + " " + member["middle_name"] + " " + member["last_name"]
-        else:
-            profile['name']= member["first_name"] + " " +  member["last_name"]
-        profile['party']= ""
-        if member["current_party"]== "R":
-            profile['party']= "Republican"
-        elif member["current_party"]== "D":
-            profile['party']= "Democrat"
-        else:
-            profile['party']= "Other"
-        roles_list= member["roles"]
-        roles= roles_list[0]
-        profile['title']= "U.S. " + roles["chamber"] + ", " + roles["title"]
-        profile['state']= roles["state"]
-        profile['service_span']= "From " + roles["start_date"]+ " to " + roles["end_date"]
-        profile['website']= member["url"]
-        if(member["twitter_account"]):
-            profile['twitter']= "@" + member["twitter_account"]
-        profile['facebook']= member["facebook_account"]
-        profile['address']= roles["office"]
-        profile['phone']= roles["phone"]
-        committees=[]
-        for committee in roles["committees"]:
-            committees.append(committee["name"])
-        profile['committees']=committees
-        return(profile)
+class CongressRep(Politician):
+    def __init__(self, name):
+        super(CongressRep, self).__init__(name, house_rep)
 
-class ExecBranchAbstract(object):
-    __metaclass__ = abc.ABCMeta
+class SenateRep(Politician):
+    def __init__(self, name):
+        super(SenateRep, self).__init__(name, exec_branch)
 
-    @abc.abstractmethod
-    def construct_profile(self):
-
-class ExecBranchRep(ExecBranchAbstract):
-    def construct_profile(self):
-        execs= requests.get("https://theunitedstates.io/congress-legislators/executive.json").json()
-        name= self._name.split(" ")
-        president={}
-        profile=self._profile
-        check= ""
-        for pres in execs:
-            pname= pres["name"]
-            first_name= pname["first"]
-            last_name= pname["last"]
-            if(first_name== name[0] and last_name== name[1]):
-                profile= self._profile
-                president=pres
-                full_name= ""
-                fname= president['name']
-                for key in fname:
-                    if(key== 'nickname'):
-                        profile['nickname']= fname[key]
-                    else:
-                        full_name= full_name + fname[key] + " "
-                profile['name']= full_name
-                profile['birthday']= president['bio']['birthday']
-                profile['service_span']= []
-                for term in president['terms']:
-                    if(term['type']=="prez"):
-                        profile['title']= "President of the United States"
-                    else:
-                        profile['title']= "Vice President of the United States"
-                    term_time= "From "+ term["start"] + " to " + term['end']
-                    profile['service_span'].append(term_time)
-                    profile['party']= term['party']
-                return profile
-        
-        
+def main(name):
+    politician= SenateRep(name)
+    member_ID= politician.strategy.find_role()
+    if member_ID== 0:
+        politician= CongressRep()
+        member_ID= politician.strategy.find_role()
+    if member_ID!=0:
+        profile= politician.strategy.construct_profile()
+    politician.strategy= exec_branch
+    check= politician.strategy.find_role()
+    if check==1:
+        politician= President()
+        profile= politician.strategy.construct_profile()
+    politician.retrieve_info()
